requirements.txt
astroid==2.1.0
autopep8==1.4.4
cairocffi==1.0.2
CairoSVG==2.3.1
certifi==2018.11.29
cffi==1.12.3
chardet==3.0.4
cssselect2==0.2.1
defusedxml==0.5.0
diff-match-patch==20121119
Django==2.1.3
django-adminactions==1.5
django-appconf==1.0.2
django-bower==5.2.0
django-classy-tags==0.8.0
django-constance==2.4.0
django-crispy-forms==1.7.2
django-easy-pdf==0.1.1
django-extensions==2.1.4
django-grappelli==2.12.1
django-grappelli-extras==1.2.0
django-image-cropping==1.2.0
django-import-export==1.1.0
django-picklefield==2.0
django-simple-history==2.7.3
django-user-agents==0.4.0
easy-thumbnails==2.5
et-xmlfile==1.0.1
gender-guesser==0.4.0
html5lib==1.0.1
httplib2==0.12.0
idna==2.8
isort==4.3.4
jdcal==1.4
lazy-object-proxy==1.3.1
mccabe==0.6.1
numpy==1.17.0
oauthlib==3.0.0
odfpy==1.3.6
openpyxl==2.5.9
pandas==0.24.2
pep8==1.7.1
Pillow==5.3.0
psycopg2==2.7.6.1
pycodestyle==2.5.0
pycparser==2.19
pycrypto==2.6.1
pydot==1.4.0
PyJWT==1.7.1
pylint==2.2.2
pyparsing==2.3.0
PyPDF2==1.26.0
Pyphen==0.9.5
python-dateutil==2.8.0
python3-openid==3.1.0
pytz==2018.7
PyYAML==3.13
redis==3.2.1
reportlab==3.5.12
requests==2.21.0
requests-oauthlib==1.2.0
six==1.11.0
social-auth-app-django==3.1.0
social-auth-core==3.0.0
tablib==0.12.1
tinycss2==1.0.2
typed-ast==1.1.0
ua-parser==0.8.0
unicodecsv==0.14.1
urllib3==1.24.1
user-agents==2.0
WeasyPrint==47
webencodings==0.5.1
wrapt==1.10.11
xhtml2pdf==0.2.3
xlrd==1.1.0
xlwt==1.3.0
